requirements.txt
# Web框架
Flask==3.0.0
Werkzeug==3.0.1

# YAML处理
PyYAML==6.0.1

# HTTP请求
requests==2.31.0

# 安全增强
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# 输入验证
cerberus==1.3.5

# 缓存支持
Flask-Caching==2.1.0

# 开发和测试工具（可选）
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# 类型检查（可选）
mypy>=1.0.0
types-PyYAML>=6.0.0
types-requests>=2.0.0 